+++ backend/src/modules/recommendations/sessions.rs
@@ -71,7 +71,7 @@ impl RecommendationsService {
             "SELECT sc_track_id FROM user_events
              WHERE sc_user_id = $1
                AND event_type IN ('full_play', 'like', 'local_like', 'playlist_add')
-               AND created_at > NOW() - make_interval(weeks => $2)
+               AND created_at > NOW() - make_interval(weeks => $2::int)
                AND ABS(EXTRACT(HOUR FROM created_at)::int - $3) <= $4
                AND EXTRACT(DOW FROM created_at)::int = $5
              ORDER BY created_at DESC
@@ -137,7 +137,7 @@ async fn recent_played_ids(
         "SELECT sc_track_id FROM user_events
          WHERE sc_user_id = $1
            AND event_type IN ('full_play', 'like', 'local_like', 'playlist_add')
-           AND created_at > NOW() - make_interval(hours => $2)
+           AND created_at > NOW() - make_interval(hours => $2::int)
          ORDER BY created_at DESC
          LIMIT $3",
     )
